requirements.txt
flask==3.0.3
flask-cors==4.0.1
pillow>=10.4.0
numpy>=1.26.4
requests==2.32.3
scikit-learn>=1.5.2
torch==2.3.0
torchvision==0.18.0
tensorflow==2.16.1
tensorflow-hub==0.16.1
gunicorn==22.0.0